task: Automatize filter options based on the previous field choice
The region field determines the options in the province field and this
determines the municipio options.

The regions, provinces and fuel types are stored on a .js file loaded by
the client. Maybe storing the data on Mongo db and getting it through
axios would be more optimized. The municipios are retrieved from an
external API @ geoportalgasolineras.es

The automatization for filling up the selection tags could be refactored
a lot, since the functionallity is pretty much the same for every field